refactor(react-ui-board): unify Grid into Board (Board unification phase 2a)
Rename the engine-driven Grid composite to Board and delete the old free-placement
Board, per DESIGN.md. The engine (grid engine) keeps its type names; only the
component is renamed (the Grid name is reserved for the lit data-table).

- Grid/{engine,geometry,Board,BoardCell,stories} moved into Board/; component
  symbols renamed Grid*→Board* (Board.Root/Container/Viewport/Content/Backdrop/Cell,
  useBoardContext, BoardController, BoardRootProps, BoardCellProps).
- Old Board (centre-origin free canvas, zoom, toolbar) removed; persistence schema
  (BoardLayout {size, cells}, defaultLayout, Position/Size/CellLayout) retained so
  plugin-board's ECHO type is unchanged.
- Rewire plugin-board BoardArticle to the unified API: maps persisted {size, cells}
  to the component's {items} layout + bounds, normalizes legacy signed/centre coords
  to 0-based, drops on change via onChange, resolver=resizeToFit (no overlaps).

Deferred to phase 2b (noted with TODOs): zoom/overview + Board.Toolbar + object
picker, and the pluggable coordinate-origin projection.
