+++ packages/plugins/plugin-board/src/containers/BoardArticle/BoardArticle.tsx
@@ -4,35 +4,39 @@
 
 import { useAtomValue } from '@effect-atom/atom-react';
 import * as Atom from '@effect-atom/atom/Atom';
-import React, { useCallback, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useMemo, useRef } from 'react';
 
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface } from '@dxos/app-toolkit/ui';
-import { Filter, Obj, Ref } from '@dxos/echo';
+import { Obj, Ref } from '@dxos/echo';
 import { useObject } from '@dxos/echo-react';
 import { invariant } from '@dxos/invariant';
 import { EID } from '@dxos/keys';
 import { Markdown } from '@dxos/plugin-markdown';
-import { useQuery } from '@dxos/react-client/echo';
 import { Panel } from '@dxos/react-ui';
 import { useAttention } from '@dxos/react-ui-attention';
-import { Board, type BoardController, type BoardRootProps, type Position } from '@dxos/react-ui-board';
-import { ObjectPicker, type ObjectPickerContentProps } from '@dxos/react-ui-form';
-import { isNonNullable } from '@dxos/util';
+import { Board, type BoardController, type BoardRootProps, type Layout, resizeToFit } from '@dxos/react-ui-board';
 
 import { type Board as BoardType } from '#types';
 
-const DEFAULT_POSITION = { x: 0, y: 0 } satisfies Position;
-
-type PickerState = {
-  position: Position;
-};
-
 export type BoardArticleProps = AppSurface.ObjectArticleProps<BoardType.Board>;
 
+// Legacy boards stored cells with a centre origin (signed coords); the engine is 0-based, so shift all
+// cells to non-negative before handing them to the component. Freshly-created boards are already 0-based.
+const normalizeCells = (cells: BoardType.Board['layout']['cells']): Layout['items'] => {
+  const values = Object.values(cells);
+  const minX = Math.min(0, ...values.map((cell) => cell.x));
+  const minY = Math.min(0, ...values.map((cell) => cell.y));
+  if (minX === 0 && minY === 0) {
+    return cells;
+  }
+  return Object.fromEntries(
+    Object.entries(cells).map(([id, cell]) => [id, { ...cell, x: cell.x - minX, y: cell.y - minY }]),
+  );
+};
+
 export const BoardArticle = ({ role, subject: board, attendableId }: BoardArticleProps) => {
-  const { hasAttention } = useAttention(attendableId);
-  const db = Obj.getDatabase(board);
+  const { hasAttention: _hasAttention } = useAttention(attendableId);
   const [boardItems] = useObject(board, 'items');
   const itemsAtom = useMemo(
     () =>
@@ -51,47 +55,31 @@ export const BoardArticle = ({ role, subject: board, attendableId }: BoardArticl
   const items = useAtomValue(itemsAtom);
 
   const controller = useRef<BoardController>(null);
-  const addTriggerRef = useRef<HTMLButtonElement | null>(null);
-  const [pickerState, setPickerState] = useState<PickerState | null>(null);
 
-  // TODO(burdon): Use search.
-  const objects = useQuery(db, Filter.everything());
-  const options = useMemo<ObjectPickerContentProps['options']>(
-    () =>
-      objects
-        .filter((obj) => obj.id !== board.id)
-        .map((obj) => {
-          const label = Obj.getLabel(obj);
-          if (label) {
-            return {
-              id: obj.id,
-              label,
-              hue: 'neutral' as const,
-            };
-          }
-        })
-        .filter(isNonNullable)
-        .sort(({ label: a }, { label: b }) => a.toLocaleLowerCase().localeCompare(b.toLocaleLowerCase())),
-    [objects],
+  const layout = useMemo<Layout>(() => ({ items: normalizeCells(board.layout.cells) }), [board.layout.cells]);
+  const bounds = useMemo(
+    () => ({ columns: board.layout.size.width, rows: board.layout.size.height }),
+    [board.layout.size.width, board.layout.size.height],
+  );
+
+  const handleChange = useCallback<NonNullable<BoardRootProps['onChange']>>(
+    (next) => {
+      Obj.update(board, (board) => {
+        board.layout.cells = next.items;
+      });
+    },
+    [board],
   );
 
   const handleAdd = useCallback<NonNullable<BoardRootProps['onAdd']>>(
-    async (anchor, position) => {
+    (position) => {
       const db = Obj.getDatabase(board);
       invariant(db);
-      // Grid backdrop "+" supplies a position → create a new Markdown document directly.
-      // Toolbar "+" omits position → fall back to the picker over existing objects.
-      if (position) {
-        const doc = db.add(Markdown.make());
-        Obj.update(board, (board) => {
-          board.items.push(Ref.make(doc));
-          board.layout.cells[doc.id.toString()] = position;
-        });
-        return;
-      }
-
-      addTriggerRef.current = anchor;
-      setPickerState({ position: DEFAULT_POSITION });
+      const doc = db.add(Markdown.make());
+      Obj.update(board, (board) => {
+        board.items.push(Ref.make(doc));
+        board.layout.cells[doc.id.toString()] = position;
+      });
     },
     [board],
   );
@@ -114,77 +102,41 @@ export const BoardArticle = ({ role, subject: board, attendableId }: BoardArticl
     [board],
   );
 
-  const handleMove = useCallback<NonNullable<BoardRootProps['onMove']>>(
-    (id, position) => {
-      const layout = board.layout.cells[id];
-      Obj.update(board, (board) => {
-        board.layout.cells[id] = { ...layout, ...position };
-      });
-    },
-    [board],
-  );
-
-  const handleSelect = useCallback<NonNullable<ObjectPickerContentProps['onSelect']>>(
-    (id) => {
-      if (!pickerState) {
-        return;
-      }
-
-      // Find the selected object by id from the space.
-      const selectedObject = objects.find((obj) => obj.id === id);
-      if (!Obj.isObject(selectedObject)) {
-        return;
-      }
-
-      // Create a reference to the selected object and add it to the board.
-      Obj.update(board, (board) => {
-        board.items.push(Ref.make(selectedObject));
-
-        // Set the layout position for the new item.
-        board.layout.cells[selectedObject.id.toString()] = pickerState.position;
-      });
-
-      // Close the picker.
-      setPickerState(null);
-    },
-    [pickerState, objects, board],
-  );
-
   return (
-    <Board.Root ref={controller} layout={board.layout} onAdd={handleAdd} onDelete={handleDelete} onMove={handleMove}>
-      <ObjectPicker.Root
-        open={!!pickerState}
-        onOpenChange={(nextOpen: boolean) => {
-          setPickerState(nextOpen ? { position: DEFAULT_POSITION } : null);
-        }}
-      >
-        <Panel.Root role={role}>
-          <Panel.Toolbar asChild>
-            <Board.Toolbar disabled={!hasAttention} />
-          </Panel.Toolbar>
-          <Panel.Content asChild>
-            <Board.Container>
-              <Board.Viewport classNames='border-none'>
-                <Board.Backdrop />
-                <Board.Content>
-                  {items?.map((item, index) => (
-                    <Board.Cell item={item} key={index} layout={board.layout?.cells[item.id] ?? { x: 0, y: 0 }}>
+    <Panel.Root role={role}>
+      <Panel.Content asChild>
+        <Board.Root
+          ref={controller}
+          layout={layout}
+          bounds={bounds}
+          mode='float'
+          resolver={resizeToFit}
+          onChange={handleChange}
+          onAdd={handleAdd}
+          onDelete={handleDelete}
+        >
+          <Board.Container classNames='absolute inset-0'>
+            <Board.Viewport>
+              <Board.Backdrop />
+              <Board.Content>
+                {items?.map((item) => {
+                  const itemLayout = layout.items[item.id];
+                  return itemLayout ? (
+                    <Board.Cell item={item} key={item.id} layout={itemLayout}>
                       <Surface.Surface
                         type={AppSurface.CardContent}
                         data={{ subject: item, editable: true }}
                         limit={1}
                       />
                     </Board.Cell>
-                  ))}
-                </Board.Content>
-              </Board.Viewport>
-            </Board.Container>
-          </Panel.Content>
-        </Panel.Root>
-        <ObjectPicker.Content options={options} onSelect={handleSelect} classNames='dx-card-popover-width' />
-        <ObjectPicker.VirtualTrigger virtualRef={addTriggerRef} />
-      </ObjectPicker.Root>
-    </Board.Root>
+                  ) : null;
+                })}
+              </Board.Content>
+            </Board.Viewport>
+          </Board.Container>
+        </Board.Root>
+      </Panel.Content>
+    </Panel.Root>
   );
 };
