+++ packages/ui/react-ui-board/src/components/Board/Board.tsx
@@ -1,15 +1,13 @@
 //
-// Copyright 2025 DXOS.org
+// Copyright 2026 DXOS.org
 //
 
-import { autoScrollForElements } from '@atlaskit/pragmatic-drag-and-drop-auto-scroll/element';
-import { combine } from '@atlaskit/pragmatic-drag-and-drop/combine';
 import { dropTargetForElements } from '@atlaskit/pragmatic-drag-and-drop/element/adapter';
 import { composeRefs } from '@radix-ui/react-compose-refs';
 import { createContext } from '@radix-ui/react-context';
 import React, {
   type ComponentPropsWithoutRef,
-  type MouseEvent,
+  type MutableRefObject,
   type PropsWithChildren,
   forwardRef,
   useCallback,
@@ -19,38 +17,35 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { useResizeDetector } from 'react-resize-detector';
 
 import { invariant } from '@dxos/invariant';
+import { IconButton, ScrollArea, type ThemedClassName, usePx, useTranslation } from '@dxos/react-ui';
 import {
-  IconButton,
-  type ThemedClassName,
-  Toolbar,
-  type ToolbarRootProps,
-  usePx,
-  useTranslation,
-} from '@dxos/react-ui';
-import { composable, composableProps } from '@dxos/react-ui';
-import { mx } from '@dxos/ui-theme';
+  type DndContainerHandler,
+  type DndPlaceholderData,
+  getSourceData,
+  useContainerId,
+  useDndRootContext,
+} from '@dxos/react-ui-dnd';
+import { cardDefaultInlineSize, mx } from '@dxos/ui-theme';
 
 import { translationKey } from '#translations';
 
 import { BoardCell, type BoardCellProps } from './BoardCell';
-import { defaultGrid, defaultLayout } from './defs';
-import { type BoardGeometry, type Rect, getBoardBounds, getBoardRect, getCenter } from './geometry';
-import { type BoardLayout, type Position, type Size } from './types';
+import {
+  type Bounds,
+  type DropResolver,
+  type GridConstraints,
+  type GridMode,
+  type GridPosition,
+  type Layout,
+  pushToFit,
+} from './engine';
+import { type GridCellSize, type Rect, cellRect, getColumnCount, getRowCount, gridBounds } from './geometry';
 
-// TODO(burdon): Infinite canvas: hierarchical zoom.
-// TODO(burdon): Drag handles to resize.
-// TODO(burdon): Synthetic scrollbars.
-// TODO(burdon): Prevent browser nav when scrolling to edge.
-
-interface BoardController {
-  /** Center the board on the given cell or position. */
-  center: (cell?: string | Position) => void;
-  /** Toggle zoom mode. */
-  toggleZoom: () => void;
-}
+// TODO(burdon): Multi-select + keyboard move/resize.
+// TODO(burdon): Zoom/overview + toolbar (port from the previous Board — deferred).
+// TODO(burdon): Pluggable coordinate origin (top-left vs centre) as a projection — deferred.
 
 //
 // Context
@@ -58,90 +53,238 @@ interface BoardController {
 
 type BoardContextValue = {
   readonly: boolean;
-  layout: BoardLayout;
-  grid: BoardGeometry;
-  bounds: Size;
-  center: Position;
-  zoom: boolean;
-  controller: BoardController;
-  onSelect?: (id: string) => void;
+  layout: Layout;
+  mode: GridMode;
+  /** Cell size and gap in px (converted from the `cellSize`/`gap` props, which are in rem). */
+  cellSize: GridCellSize;
+  gap: number;
+  /** Column/row extent to render (the backdrop shows at least this; grows with content). */
+  columns: number;
+  rows: number;
+  containerId: string;
+  /** During an active drag, the layout the board would settle into — tiles animate to these
+   * positions and spring back to `layout` when the drag ends without a drop. Undefined when idle. */
+  previewLayout?: Layout;
+  /** True while a tile is being resized (a pointer drag, not a Dnd drag); gates the resize auto-scroll. */
+  resizing: boolean;
+  /** Scroll viewport element; set by `Board.Container`, used by the controller to center. */
+  viewportRef: MutableRefObject<HTMLDivElement | null>;
+  /** Scroll the viewport so the board is centered. */
+  center: () => void;
+  onAdd?: (position: GridPosition) => void;
   onDelete?: (id: string) => void;
-  onMove?: (id: string, position: Position) => void;
-  onAdd?: (anchor: HTMLButtonElement, position?: Position) => void;
+  onResize: (id: string, size: { w: number; h: number }, constraints?: GridConstraints) => void;
+  /** Report an in-progress resize (snapped cells) so the engine runs live and other tiles move;
+   * pass null on drop/cancel. */
+  onResizePreview: (id: string, size: { w: number; h: number } | null) => void;
 };
 
 const [BoardContextProvider, useBoardContext] = createContext<BoardContextValue>('BoardContext');
 
+/**
+ * Imperative handle exposed by `Board.Root` via ref.
+ */
+export type BoardController = {
+  /** Scroll the viewport so the board content is centered. */
+  center: () => void;
+};
+
 //
 // Root
-// NOTE: The Root is headless, which allows the Controls and Container to be in different subtrees.
+// NOTE: Headless (renders no DOM); registers a single Dnd container handler for the whole board.
 //
 
-type BoardRootProps = PropsWithChildren<
-  Partial<Pick<BoardContextValue, 'readonly' | 'layout' | 'grid' | 'onSelect' | 'onDelete' | 'onMove' | 'onAdd'>>
->;
+const BOARD_ROOT_NAME = 'Board.Root';
+
+type BoardRootProps = PropsWithChildren<{
+  layout: Layout;
+  mode?: GridMode;
+  /**
+   * Board extent in cells: `columns` bounds the horizontal axis (clamp + right-push fallback);
+   * `rows` is a minimum (the backdrop shows at least this, and the board grows past it). Omit
+   * `columns` for a board whose width is derived from its content.
+   */
+  bounds?: Bounds;
+  /**
+   * Strategy applied when a tile is dropped or resized. Returns the resulting layout, or `null` to
+   * reject (the tile springs back). Defaults to {@link pushToFit} (push occupants out of the way).
+   */
+  resolver?: DropResolver;
+  /** Cell size in rem. */
+  cellSize?: GridCellSize;
+  /** Gap between cells in rem. */
+  gap?: number;
+  /**
+   * Milliseconds the drag must dwell on a target cell before the resolver runs against it. While
+   * sweeping across cells faster than this, the other tiles stay put — so a tile can be dragged
+   * over/past existing ones and only displaces them once the drag settles. 0 = immediate.
+   */
+  settleDelay?: number;
+  readonly?: boolean;
+  onChange?: (layout: Layout) => void;
+  onAdd?: (position: GridPosition) => void;
+  onDelete?: (id: string) => void;
+}>;
+
+// Default to a compact cell (~half a default card) so a board fits more tiles on screen; consumers
+// can pass a larger `cellSize` for a card-sized board.
+const defaultCellSize: GridCellSize = { width: cardDefaultInlineSize / 2, height: cardDefaultInlineSize / 2 };
+const defaultGap = 1;
 
 const BoardRoot = forwardRef<BoardController, BoardRootProps>(
   (
-    { children, readonly, layout = defaultLayout, grid = defaultGrid, onSelect, onDelete, onMove, onAdd },
+    {
+      children,
+      layout,
+      mode = 'pack',
+      bounds,
+      resolver = pushToFit,
+      cellSize = defaultCellSize,
+      gap = defaultGap,
+      settleDelay = 500,
+      readonly,
+      onChange,
+      onAdd,
+      onDelete,
+    },
     forwardedRef,
   ) => {
     const remInPx = usePx(1);
-    const gridInPx = useMemo(() => {
-      return {
-        size: { width: grid.size.width * remInPx, height: grid.size.height * remInPx },
-        gap: grid.gap * remInPx,
-        ...(grid?.overScroll && { overScroll: grid.overScroll * remInPx }),
-      };
-    }, [remInPx, grid]);
-    const bounds = useMemo<Size>(() => getBoardBounds(layout.size, gridInPx), [layout, gridInPx]);
+    const cellSizePx = useMemo<GridCellSize>(
+      () => ({ width: cellSize.width * remInPx, height: cellSize.height * remInPx }),
+      [remInPx, cellSize.width, cellSize.height],
+    );
+    const gapPx = gap * remInPx;
+    const columns = useMemo(() => bounds?.columns ?? getColumnCount(layout), [bounds?.columns, layout]);
+    const rows = useMemo(() => Math.max(getRowCount(layout), bounds?.rows ?? 0), [layout, bounds?.rows]);
 
-    const [zoom, setZoom] = useState(false);
-    const [center, setCenter] = useState({ x: bounds.width / 2, y: bounds.height / 2 });
+    // Apply the resolver with the board's bounds/mode; returns the next layout or null (reject).
+    const resolve = useCallback(
+      (id: string, to: GridPosition, constraints?: GridConstraints): Layout | null =>
+        resolver(layout, id, to, { bounds, constraints, mode }),
+      [resolver, layout, bounds, mode],
+    );
 
-    // External controller.
-    const controller = useMemo<BoardController>(
-      () => ({
-        center: (cell) => {
-          if (cell) {
-            const position = typeof cell === 'string' ? layout?.cells[cell] : cell;
-            if (position) {
-              const center = getCenter(getBoardRect(gridInPx, position));
-              setCenter({ x: bounds.width / 2 + center.x, y: bounds.height / 2 + center.y });
-              setZoom(false);
-            }
-          } else {
-            setCenter({ x: bounds.width / 2, y: bounds.height / 2 });
+    const containerId = useContainerId('board');
+
+    // Scroll viewport (set by Board.Container) + imperative centering, exposed via the Root ref.
+    const viewportRef = useRef<HTMLDivElement | null>(null);
+    const center = useCallback(() => {
+      const el = viewportRef.current;
+      if (el) {
+        el.scrollTo({ left: (el.scrollWidth - el.clientWidth) / 2, top: (el.scrollHeight - el.clientHeight) / 2 });
+      }
+    }, []);
+    useImperativeHandle(forwardedRef, () => ({ center }), [center]);
+
+    const onResize = useCallback(
+      (id: string, size: { w: number; h: number }, constraints?: GridConstraints) => {
+        const current = layout.items[id];
+        if (!current) {
+          return;
+        }
+        // Resize keeps x/y and applies the new span through the same resolver as a move.
+        const next = resolve(id, { ...current, w: size.w, h: size.h }, constraints);
+        if (next) {
+          onChange?.(next);
+        }
+      },
+      [layout, resolve, onChange],
+    );
+
+    // Register this board's container handler. Re-registers (by the stable `containerId` key) whenever
+    // `resolve`/`onChange` change so `onDrop` always closes over the current layout.
+    // NOTE: `addContainer`/`removeContainer` are re-created every `Dnd.Root` render; they are
+    // intentionally left out of the deps below (mirroring `Mosaic.Container`) so this effect only
+    // reruns when the board's own state changes, not on every drag frame.
+    const { addContainer, removeContainer, dragging } = useDndRootContext(BOARD_ROOT_NAME);
+
+    // In-progress resize (snapped cells), reported by the resizing cell; drives a live preview.
+    const [resizePreview, setResizePreview] = useState<{ id: string; w: number; h: number } | null>(null);
+    const onResizePreview = useCallback((id: string, size: { w: number; h: number } | null) => {
+      setResizePreview(size ? { id, ...size } : null);
+    }, []);
+
+    // The cell the drag is currently over (this board's own tile, over a placeholder), or undefined.
+    const dragTarget =
+      dragging && dragging.source.data.containerId === containerId && dragging.target?.data.type === 'placeholder'
+        ? dragging.target.data.location
+        : undefined;
+
+    // Debounce the drag target: only adopt it as the "settled" cell after the drag has dwelt on it for
+    // `settleDelay` ms. Sweeping across cells faster than that keeps the last settled cell, so the
+    // other tiles don't scatter while the tile passes over them (and it can reach the far side).
+    const [settledTarget, setSettledTarget] = useState<{ x: number; y: number } | undefined>(undefined);
+    useEffect(() => {
+      if (!dragTarget || settleDelay <= 0) {
+        setSettledTarget(dragTarget);
+        return;
+      }
+      const timer = setTimeout(() => setSettledTarget(dragTarget), settleDelay);
+      return () => clearTimeout(timer);
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [dragTarget?.x, dragTarget?.y, settleDelay]);
+
+    // The layout the board would settle into during an in-progress move or resize, so the other tiles
+    // animate out of the way (and spring back to `layout` when the gesture ends). Resize takes
+    // precedence (it isn't a Dnd drag, so `dragging` is unset during it); the move uses the settled
+    // (debounced) target so pushes only happen once the drag pauses on a cell.
+    const previewLayout = useMemo<Layout | undefined>(() => {
+      if (resizePreview) {
+        const current = layout.items[resizePreview.id];
+        return current
+          ? (resolve(resizePreview.id, { ...current, w: resizePreview.w, h: resizePreview.h }) ?? undefined)
+          : undefined;
+      }
+      if (dragging && dragging.source.data.containerId === containerId && settledTarget) {
+        const id = dragging.source.data.id;
+        const current = layout.items[id];
+        return current ? (resolve(id, { ...current, x: settledTarget.x, y: settledTarget.y }) ?? undefined) : undefined;
+      }
+      return undefined;
+    }, [resizePreview, dragging, containerId, settledTarget, layout, resolve]);
+
+    useEffect(() => {
+      const handler: DndContainerHandler = {
+        id: containerId,
+        canDrop: ({ source }) => source.containerId === containerId,
+        onDrop: ({ source, target }) => {
+          if (target?.type !== 'placeholder') {
+            return;
+          }
+          const current = layout.items[source.id];
+          if (!current) {
+            return;
+          }
+          const next = resolve(source.id, { ...current, x: target.location.x, y: target.location.y });
+          if (next) {
+            onChange?.(next);
           }
         },
-        toggleZoom: () => {
-          setZoom((prev) => !prev);
-        },
-      }),
-      [layout, gridInPx, bounds],
-    );
-    useImperativeHandle(forwardedRef, () => controller, [controller]);
-
-    const handleSelect = useCallback<NonNullable<BoardContextValue['onSelect']>>(
-      (id) => {
-        controller.center(id);
-      },
-      [controller],
-    );
+      };
+      addContainer(handler);
+      return () => removeContainer(containerId);
+      // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [containerId, resolve, onChange]);
 
     return (
       <BoardContextProvider
         readonly={readonly ?? false}
         layout={layout}
-        grid={gridInPx}
-        bounds={bounds}
+        mode={mode}
+        cellSize={cellSizePx}
+        gap={gapPx}
+        columns={columns}
+        rows={rows}
+        containerId={containerId}
+        previewLayout={previewLayout}
+        resizing={!!resizePreview}
+        viewportRef={viewportRef}
         center={center}
-        zoom={zoom}
-        controller={controller}
-        onSelect={onSelect ?? handleSelect}
-        onDelete={onDelete}
-        onMove={onMove}
         onAdd={readonly ? undefined : onAdd}
+        onDelete={readonly ? undefined : onDelete}
+        onResize={onResize}
+        onResizePreview={onResizePreview}
       >
         {children}
       </BoardContextProvider>
@@ -149,69 +292,7 @@ const BoardRoot = forwardRef<BoardController, BoardRootProps>(
   },
 );
 
-BoardRoot.displayName = 'Board.Root';
-
-//
-// Container
-//
-
-const BOARD_CONTAINER_NAME = 'Board.Container';
-
-type BoardContainerProps = PropsWithChildren<ComponentPropsWithoutRef<'div'>>;
-
-const BoardContainer = composable<HTMLDivElement, BoardContainerProps>(
-  ({ classNames, children, ...props }, forwardedRef) => {
-    const containerRef = useRef<HTMLDivElement>(null);
-    const { width, height } = useResizeDetector({ targetRef: containerRef });
-    const { bounds, grid, center } = useBoardContext(BOARD_CONTAINER_NAME);
-
-    const [mounted, setMounted] = useState(false);
-
-    // Auto-center (on mount).
-    useEffect(() => {
-      const container = containerRef.current;
-      if (container && width && height) {
-        container.scrollTo({
-          left: center.x - width / 2,
-          top: center.y - height / 2,
-          behavior: mounted ? 'smooth' : 'auto',
-        });
-
-        setMounted(true);
-      }
-    }, [center, bounds, width, height]);
-
-    // Auto-scroll.
-    useEffect(() => {
-      invariant(containerRef.current);
-      return autoScrollForElements({
-        element: containerRef.current,
-      });
-    }, []);
-
-    return (
-      <div
-        {...composableProps(props, {
-          classNames: [
-            'flex items-center justify-center overflow-auto scrollbar-none overscroll-x-contain',
-            'opacity-0 transition-opacity duration-1000',
-            mounted && 'opacity-100',
-            classNames,
-          ],
-        })}
-        ref={composeRefs(containerRef, forwardedRef)}
-        style={{
-          padding: grid.overScroll ?? 0,
-        }}
-      >
-        {/* NOTE: This ensures that the children are centered if they are smaller than the container. */}
-        <div className='max-h-full max-w-full'>{children}</div>
-      </div>
-    );
-  },
-);
-
-BoardContainer.displayName = BOARD_CONTAINER_NAME;
+BoardRoot.displayName = BOARD_ROOT_NAME;
 
 //
 // Viewport
@@ -222,26 +303,143 @@ const BOARD_VIEWPORT_NAME = 'Board.Viewport';
 type BoardViewportProps = ThemedClassName<PropsWithChildren>;
 
 const BoardViewport = ({ classNames, children }: BoardViewportProps) => {
-  const { bounds, zoom } = useBoardContext(BOARD_VIEWPORT_NAME);
+  const { cellSize, gap, columns, rows } = useBoardContext(BOARD_VIEWPORT_NAME);
+  const bounds = useMemo(() => gridBounds(columns, rows, cellSize, gap), [columns, rows, cellSize, gap]);
+
   return (
-    <div
-      className={mx(
-        'relative transition-transform duration-300 border border-separator rounded-lg',
-        zoom && 'scale-50',
-        classNames,
-      )}
-      style={{
-        width: bounds.width,
-        height: bounds.height,
-      }}
-    >
-      {/* Scrollable container. */}
-      <div className={mx('absolute top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2')}>{children}</div>
+    // `m-auto` centers the board within the (flex) scroll container when it fits, and stays fully
+    // scrollable when it overflows (unlike justify-center, which would clip the top/left). Scroll
+    // snap points live on the backdrop cells (snap-to-grid), not here (see Board.Container / Backdrop).
+    <div className={mx('relative m-auto', classNames)} style={{ width: bounds.width, height: bounds.height }}>
+      {children}
     </div>
   );
 };
 
 BoardViewport.displayName = BOARD_VIEWPORT_NAME;
+
+//
+// Container
+// NOTE: Scroll viewport; a custom edge auto-scroll fires while dragging/resizing near an edge.
+//
+
+const BOARD_CONTAINER_NAME = 'Board.Container';
+
+type BoardContainerProps = ThemedClassName<PropsWithChildren>;
+
+const BoardContainer = ({ classNames, children }: BoardContainerProps) => {
+  const { viewportRef, center, resizing } = useBoardContext(BOARD_CONTAINER_NAME);
+  const localRef = useRef<HTMLDivElement>(null);
+  const ref = composeRefs(localRef, viewportRef);
+
+  // Read the live resizing flag from a ref so the mount-once auto-scroll effect always sees it.
+  const resizingRef = useRef(resizing);
+  resizingRef.current = resizing;
+
+  // Custom edge auto-scroll: engage only within a narrow edge band and ramp speed gently, so
+  // scrolling starts near the edge (not early) and is smooth — pragmatic's `autoScrollForElements`
+  // uses a large percentage-based hitbox at a fixed speed, with no public knob to soften either.
+  // `dragover` drives it for tile moves (native HTML5 drag); `pointermove` (only while a resize is
+  // active) drives it for the resize gesture. Both are gated so idle hover/scrollbar drags don't scroll.
+  useEffect(() => {
+    const element = localRef.current;
+    invariant(element);
+
+    const edge = 88;
+    const maxSpeed = 8;
+    let velocityX = 0;
+    let velocityY = 0;
+    let frame = 0;
+
+    const stop = () => {
+      velocityX = 0;
+      velocityY = 0;
+      if (frame) {
+        cancelAnimationFrame(frame);
+        frame = 0;
+      }
+      // Restore scroll-snap (removed while auto-scrolling — see tick) so manual scrolling snaps to grid.
+      element.style.scrollSnapType = '';
+    };
+    const tick = () => {
+      if (!velocityX && !velocityY) {
+        frame = 0;
+        return;
+      }
+      // Disable scroll-snap while auto-scrolling: with snap on, each incremental scrollBy re-snaps to
+      // the next grid line and the viewport jumps a whole cell at a time instead of scrolling smoothly.
+      element.style.scrollSnapType = 'none';
+      element.scrollBy({ left: velocityX, top: velocityY });
+      frame = requestAnimationFrame(tick);
+    };
+    // Ramp from 0 at the band's inner edge to maxSpeed at the very edge.
+    const speed = (depth: number) => (Math.min(depth, edge) / edge) * maxSpeed;
+    const track = (clientX: number, clientY: number) => {
+      const rect = element.getBoundingClientRect();
+      velocityX =
+        clientX < rect.left + edge
+          ? -speed(rect.left + edge - clientX)
+          : clientX > rect.right - edge
+            ? speed(clientX - (rect.right - edge))
+            : 0;
+      velocityY =
+        clientY < rect.top + edge
+          ? -speed(rect.top + edge - clientY)
+          : clientY > rect.bottom - edge
+            ? speed(clientY - (rect.bottom - edge))
+            : 0;
+      if ((velocityX || velocityY) && !frame) {
+        frame = requestAnimationFrame(tick);
+      }
+    };
+
+    const onDragOver = (event: DragEvent) => {
+      // Mark the drag as handled so the browser's own (fast, jumpy) native autoscroll stays off and
+      // only this custom, ramped scroll runs — otherwise the two compound into large uncontrolled jumps.
+      event.preventDefault();
+      track(event.clientX, event.clientY);
+    };
+    const onPointerMove = (event: PointerEvent) => {
+      if (resizingRef.current) {
+        track(event.clientX, event.clientY);
+      }
+    };
+
+    element.addEventListener('dragover', onDragOver);
+    element.addEventListener('pointermove', onPointerMove);
+    document.addEventListener('drop', stop);
+    document.addEventListener('dragend', stop);
+    document.addEventListener('pointerup', stop);
+    document.addEventListener('pointercancel', stop);
+    return () => {
+      stop();
+      element.removeEventListener('dragover', onDragOver);
+      element.removeEventListener('pointermove', onPointerMove);
+      document.removeEventListener('drop', stop);
+      document.removeEventListener('dragend', stop);
+      document.removeEventListener('pointerup', stop);
+      document.removeEventListener('pointercancel', stop);
+    };
+  }, []);
+
+  // Center the board once on mount only. Deliberately NOT re-centering when the layout/size changes,
+  // so the viewport doesn't jump after a drag or resize (`center` is stable; runs a single time).
+  useEffect(() => {
+    center();
+  }, [center]);
+
+  return (
+    <ScrollArea.Root orientation='all' classNames={classNames}>
+      {/* `flex` so the viewport's `m-auto` centers the board; overflow scrolls both axes. `snap-*`
+          makes scrolling magnetically settle on the grid lines (snap points are the backdrop cells). */}
+      <ScrollArea.Viewport ref={ref} classNames='flex snap-both snap-proximity'>
+        {children}
+      </ScrollArea.Viewport>
+    </ScrollArea.Root>
+  );
+};
+
+BoardContainer.displayName = BOARD_CONTAINER_NAME;
 
 //
 // Content
@@ -252,7 +450,11 @@ const BOARD_CONTENT_NAME = 'Board.Content';
 type BoardContentProps = ThemedClassName<ComponentPropsWithoutRef<'div'>>;
 
 const BoardContent = ({ classNames, children, ...props }: BoardContentProps) => {
-  return <div className={mx(classNames)}>{children}</div>;
+  return (
+    <div className={mx('relative', classNames)} {...props}>
+      {children}
+    </div>
+  );
 };
 
 BoardContent.displayName = BOARD_CONTENT_NAME;
@@ -265,28 +467,29 @@ const BOARD_BACKDROP_NAME = 'Board.Backdrop';
 
 type BoardBackdropProps = {};
 
-const BoardBackdrop = (props: BoardBackdropProps) => {
-  const { grid: board, layout, onAdd } = useBoardContext(BOARD_BACKDROP_NAME);
+const BoardBackdrop = (_props: BoardBackdropProps) => {
+  const { cellSize, gap, columns, rows, containerId, readonly, onAdd } = useBoardContext(BOARD_BACKDROP_NAME);
 
   const cells = useMemo(() => {
-    const cells: { position: Position; rect: Rect }[] = [];
-    for (let x = -Math.floor(layout.size.width / 2); x <= Math.floor(layout.size.width / 2); x++) {
-      for (let y = -Math.floor(layout.size.height / 2); y <= Math.floor(layout.size.height / 2); y++) {
-        cells.push({ position: { x, y }, rect: getBoardRect(board, { x, y }) });
+    const cells: { position: { x: number; y: number }; rect: Rect }[] = [];
+    for (let y = 0; y < rows; y++) {
+      for (let x = 0; x < columns; x++) {
+        cells.push({ position: { x, y }, rect: cellRect({ x, y, w: 1, h: 1 }, cellSize, gap) });
       }
     }
 
     return cells;
-  }, [layout, board]);
+  }, [columns, rows, cellSize, gap]);
 
   return (
     <div className='absolute inset-0'>
-      {cells.map(({ position, rect }, index) => (
+      {cells.map(({ position, rect }) => (
         <BoardDropTarget
-          key={index}
+          key={`${position.x}-${position.y}`}
           position={position}
           rect={rect}
-          onAddClick={(event) => onAdd?.(event.currentTarget as HTMLButtonElement, position)}
+          containerId={containerId}
+          onAddClick={readonly ? undefined : () => onAdd?.({ x: position.x, y: position.y, w: 1, h: 1 })}
         />
       ))}
     </div>
@@ -296,46 +499,36 @@ const BoardBackdrop = (props: BoardBackdropProps) => {
 BoardBackdrop.displayName = BOARD_BACKDROP_NAME;
 
 type BoardDropTargetProps = {
-  position: Position;
+  position: { x: number; y: number };
   rect: Rect;
-  onAddClick?: (event: MouseEvent) => void;
+  containerId: string;
+  onAddClick?: () => void;
 };
 
-const BoardDropTarget = ({ position, rect, onAddClick }: BoardDropTargetProps) => {
+const BoardDropTarget = ({ position, rect, containerId, onAddClick }: BoardDropTargetProps) => {
   const { t } = useTranslation(translationKey);
 
-  const [active, setActive] = useState(false);
   const ref = useRef<HTMLDivElement | null>(null);
   useEffect(() => {
     invariant(ref.current);
-    return combine(
-      dropTargetForElements({
-        element: ref.current,
-        getData: () => ({ position }),
-        onDragEnter: () => {
-          setActive(true);
-        },
-        onDragLeave: () => {
-          setActive(false);
-        },
-        onDrop: () => {
-          setActive(false);
-        },
-      }),
-    );
-  }, []);
+    return dropTargetForElements({
+      element: ref.current,
+      getData: () => ({ type: 'placeholder', containerId, location: position }) satisfies DndPlaceholderData,
+      // Only this board's own tiles may be dropped here (same-container move; no cross-container yet).
+      // The active-target highlight is drawn as the dragged tile's full footprint outline (see BoardCell),
+      // so individual cells don't need their own ring.
+      canDrop: ({ source }) => getSourceData(source)?.containerId === containerId,
+    });
+  }, [containerId, position.x, position.y]);
 
   return (
     <div
       ref={ref}
       style={rect}
-      className={mx(
-        'group/cell absolute flex items-center justify-center border rounded-sm opacity-50',
-        active ? 'border-transparent ring ring-accent-bg' : 'border-separator border-dashed',
-      )}
+      // `snap-start` aligns each cell's leading edge to the viewport, so scrolling snaps to the grid.
+      className='group/cell absolute flex snap-start items-center justify-center rounded-sm border border-dashed border-separator opacity-50'
     >
       {onAddClick && (
-        // TODO(burdon): Make this pluggable so that the container can provide a menu trigger.
         <IconButton
           icon='ph--plus--regular'
           iconOnly
@@ -349,47 +542,6 @@ const BoardDropTarget = ({ position, rect, onAddClick }: BoardDropTargetProps) =
 };
 
 //
-// Controls
-//
-
-const BOARD_TOOLBAR_NAME = 'Board.Controls';
-
-type BoardToolbarProps = ToolbarRootProps;
-
-const BoardToolbar = composable<HTMLDivElement, BoardToolbarProps>((props, forwardedRef) => {
-  const { t } = useTranslation(translationKey);
-  const { readonly, zoom, controller, onAdd } = useBoardContext(BOARD_TOOLBAR_NAME);
-
-  // TODO(burdon): Convert to MenuProvider.
-  return (
-    <Toolbar.Root {...composableProps(props)} ref={forwardedRef}>
-      <Toolbar.IconButton
-        icon='ph--crosshair--regular'
-        iconOnly
-        label={t('move-to-center.button')}
-        onClick={() => controller.center()}
-      />
-      <Toolbar.IconButton
-        icon={zoom ? 'ph--arrows-in--regular' : 'ph--arrows-out--regular'}
-        iconOnly
-        label={t('toggle-zoom.button')}
-        onClick={() => controller.toggleZoom()}
-      />
-      {!readonly && onAdd && (
-        <Toolbar.IconButton
-          icon='ph--plus--regular'
-          iconOnly
-          label={t('add-object.button')}
-          onClick={(event) => onAdd?.(event.currentTarget as HTMLButtonElement)}
-        />
-      )}
-    </Toolbar.Root>
-  );
-});
-
-BoardToolbar.displayName = BOARD_TOOLBAR_NAME;
-
-//
 // Board
 //
 
@@ -399,7 +551,6 @@ export const Board = {
   Viewport: BoardViewport,
   Content: BoardContent,
   Backdrop: BoardBackdrop,
-  Toolbar: BoardToolbar,
   Cell: BoardCell,
 };
 
@@ -408,9 +559,7 @@ export type {
   BoardCellProps,
   BoardContainerProps,
   BoardContentProps,
-  BoardController,
   BoardRootProps,
-  BoardToolbarProps,
   BoardViewportProps,
 };
