+++ packages/ui/react-ui-board/src/components/Board/geometry.ts
@@ -1,34 +1,59 @@
 //
-// Copyright 2025 DXOS.org
+// Copyright 2026 DXOS.org
 //
-
-import { type CellLayout, type Position, type Size } from './types';
 
 export type Rect = Pick<DOMRect, 'left' | 'top' | 'width' | 'height'>;
 
-export type BoardGeometry = {
-  size: Size;
-  gap: number;
-  overScroll?: number;
-};
+/** Cell size in px (or any consistent unit, matching whatever unit `gap` is expressed in). */
+export type GridCellSize = { width: number; height: number };
 
-export const getCenter = (rect: Rect): Position => {
-  return {
-    x: rect.left + rect.width / 2,
-    y: rect.top + rect.height / 2,
-  };
-};
-
-export const getBoardBounds = (size: Size, board: BoardGeometry): Size => ({
-  width: size.width * (board.size.width + board.gap) + board.gap,
-  height: size.height * (board.size.height + board.gap) + board.gap,
+/**
+ * Pixel rect for a grid cell/item given its integer cell coordinates and size. Unlike Board (which
+ * centers cell (0,0)), Grid coordinates are 0-indexed with a top-left origin. A leading `gap` is
+ * included before the first cell so the gutter is symmetric on all four edges (matching the trailing
+ * `gap` in {@link gridBounds}); without it the grid hugs the top-left but has a margin bottom-right.
+ */
+export const cellRect = (
+  { x, y, w, h }: { x: number; y: number; w: number; h: number },
+  cellSize: GridCellSize,
+  gap: number,
+): Rect => ({
+  left: gap + x * (cellSize.width + gap),
+  top: gap + y * (cellSize.height + gap),
+  width: w * cellSize.width + Math.max(0, w - 1) * gap,
+  height: h * cellSize.height + Math.max(0, h - 1) * gap,
 });
 
-export const getBoardRect = (board: BoardGeometry, { x, y, width = 1, height = 1 }: CellLayout): Rect => {
-  return {
-    left: x * (board.size.width + board.gap) - board.size.width / 2,
-    top: y * (board.size.height + board.gap) - board.size.height / 2,
-    width: width * board.size.width + Math.max(0, width - 1) * board.gap,
-    height: height * board.size.height + Math.max(0, height - 1) * board.gap,
-  };
+/**
+ * Overall pixel size of a grid with the given column/row count.
+ */
+export const gridBounds = (
+  columns: number,
+  rows: number,
+  cellSize: GridCellSize,
+  gap: number,
+): { width: number; height: number } => ({
+  width: columns * (cellSize.width + gap) + gap,
+  height: rows * (cellSize.height + gap) + gap,
+});
+
+/** A tile position (span optional, defaulting to 1×1) as stored in a layout keyed by id. */
+type Position = { x: number; y: number; w?: number; h?: number };
+
+/**
+ * Number of rows to render: the lowest occupied row edge plus a few spare rows so there's always
+ * room to drag/add below the current content.
+ */
+export const getRowCount = (layout: { items: Record<string, Position> }, spareRows = 3): number => {
+  const maxBottom = Object.values(layout.items).reduce((bottom, item) => Math.max(bottom, item.y + (item.h ?? 1)), 0);
+  return maxBottom + spareRows;
+};
+
+/**
+ * Number of columns to render when no explicit column bound is given: the right-most occupied edge
+ * plus a couple of spare columns (used for the backdrop extent of an unbounded board).
+ */
+export const getColumnCount = (layout: { items: Record<string, Position> }, spareColumns = 2): number => {
+  const maxRight = Object.values(layout.items).reduce((right, item) => Math.max(right, item.x + (item.w ?? 1)), 0);
+  return maxRight + spareColumns;
 };
