+++ packages/ui/react-ui-board/src/components/Board/index.ts
@@ -2,7 +2,8 @@
 // Copyright 2025 DXOS.org
 //
 
-export * from './defs';
+export * from './engine';
+export * from './geometry';
 export * from './types';
 
 export * from './Board';
