--- packages/ui/react-ui-board/src/components/Grid/index.ts
@@ -1,7 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-export * from './engine';
-export * from './geometry';
-export * from './Grid';
