+++ packages/ui/react-ui-board/src/components/index.ts
@@ -3,4 +3,3 @@
 //
 
 export * from './Board';
-export * from './Grid';
